+++ gen-annotations/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod gff;
+pub mod projection;
 #[cfg(test)]
 pub mod test_helpers;
 pub mod translate;
