+++ gen-annotations/src/projection.rs
@@ -1,0 +1,219 @@
+use std::cmp::{max, min};
+
+use gen_core::{
+    HashId, Strand,
+    path::PathBlock,
+    range::{OrderedMerge, Range, merge_ordered_items},
+};
+use gen_models::accession::AccessionEdge;
+
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub struct AnnotationSegment {
+    pub node_id: HashId,
+    pub range: Range,
+    pub strand: Strand,
+}
+
+impl OrderedMerge for AnnotationSegment {
+    fn should_merge_with(&self, next: &Self) -> bool {
+        self.strand == next.strand
+            && next.range.start >= self.range.start
+            && next.range.start <= self.range.end
+    }
+
+    fn merge_with(&mut self, next: &Self) {
+        self.range.end = max(self.range.end, next.range.end);
+    }
+}
+
+pub fn accession_edges_to_segments(edges: &[AccessionEdge]) -> Vec<AnnotationSegment> {
+    let mut segments = Vec::new();
+    let mut current_node = None;
+    let mut current_start = None;
+    let mut current_strand = None;
+
+    for edge in edges {
+        if edge.source_coordinate < 0 {
+            current_node = Some(edge.target_node_id);
+            current_start = Some(edge.target_coordinate);
+            current_strand = Some(edge.target_strand);
+            continue;
+        }
+
+        if let (Some(node_id), Some(start), Some(strand)) =
+            (current_node, current_start, current_strand)
+        {
+            let (segment_start, segment_end) = if start <= edge.source_coordinate {
+                (start, edge.source_coordinate)
+            } else {
+                (edge.source_coordinate, start)
+            };
+            if segment_end > segment_start {
+                segments.push(AnnotationSegment {
+                    node_id,
+                    range: Range {
+                        start: segment_start,
+                        end: segment_end,
+                    },
+                    strand,
+                });
+            }
+        }
+
+        if edge.target_coordinate < 0 {
+            break;
+        }
+
+        current_node = Some(edge.target_node_id);
+        current_start = Some(edge.target_coordinate);
+        current_strand = Some(edge.target_strand);
+    }
+
+    segments
+}
+
+pub fn merge_annotation_segments(segments: Vec<AnnotationSegment>) -> Vec<AnnotationSegment> {
+    merge_ordered_items(
+        segments
+            .into_iter()
+            .filter(|segment| segment.range.end > segment.range.start)
+            .collect(),
+    )
+}
+
+pub fn project_single_annotation_segment(
+    segment: &AnnotationSegment,
+    path_blocks: &[PathBlock],
+) -> Vec<AnnotationSegment> {
+    merge_annotation_segments(
+        path_blocks
+            .iter()
+            .filter_map(|block| {
+                if block.node_id != segment.node_id {
+                    return None;
+                }
+                let overlap_start = max(segment.range.start, block.sequence_start);
+                let overlap_end = min(segment.range.end, block.sequence_end);
+                if overlap_end <= overlap_start {
+                    return None;
+                }
+
+                let (start, end) = if block.strand == Strand::Reverse {
+                    (
+                        block.path_start + (block.sequence_end - overlap_end),
+                        block.path_start + (block.sequence_end - overlap_start),
+                    )
+                } else {
+                    (
+                        block.path_start + (overlap_start - block.sequence_start),
+                        block.path_start + (overlap_end - block.sequence_start),
+                    )
+                };
+                let strand = if block.strand == Strand::Reverse {
+                    segment.strand.complement()
+                } else {
+                    segment.strand
+                };
+
+                Some(AnnotationSegment {
+                    node_id: block.node_id,
+                    range: Range { start, end },
+                    strand,
+                })
+            })
+            .collect(),
+    )
+}
+
+pub fn project_annotation_segments(
+    accession_segments: &[AnnotationSegment],
+    path_blocks: &[PathBlock],
+    preserve_part_boundaries: bool,
+) -> Vec<AnnotationSegment> {
+    let projected = accession_segments
+        .iter()
+        .flat_map(|segment| project_single_annotation_segment(segment, path_blocks))
+        .collect::<Vec<_>>();
+
+    if preserve_part_boundaries {
+        projected
+    } else {
+        merge_annotation_segments(projected)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use gen_core::{HashId, path::PathBlock};
+
+    use super::*;
+
+    fn block(
+        node_id: &str,
+        sequence_start: i64,
+        sequence_end: i64,
+        path_start: i64,
+        strand: Strand,
+    ) -> PathBlock {
+        PathBlock {
+            node_id: HashId::convert_str(node_id),
+            block_sequence: String::new(),
+            sequence_start,
+            sequence_end,
+            path_start,
+            path_end: path_start + (sequence_end - sequence_start),
+            strand,
+        }
+    }
+
+    #[test]
+    fn test_merge_annotation_segments_merges_overlapping_ranges() {
+        let node_id = HashId::convert_str("node");
+        let merged = merge_annotation_segments(vec![
+            AnnotationSegment {
+                node_id,
+                range: Range { start: 2, end: 5 },
+                strand: Strand::Forward,
+            },
+            AnnotationSegment {
+                node_id,
+                range: Range { start: 4, end: 8 },
+                strand: Strand::Forward,
+            },
+        ]);
+
+        assert_eq!(
+            merged,
+            vec![AnnotationSegment {
+                node_id,
+                range: Range { start: 2, end: 8 },
+                strand: Strand::Forward,
+            }]
+        );
+    }
+
+    #[test]
+    fn test_project_single_annotation_segment_complements_reverse_blocks() {
+        let node_id = HashId::convert_str("node");
+        let projected = project_single_annotation_segment(
+            &AnnotationSegment {
+                node_id,
+                range: Range { start: 10, end: 20 },
+                strand: Strand::Forward,
+            },
+            &[block("node", 0, 30, 100, Strand::Reverse)],
+        );
+
+        assert_eq!(
+            projected,
+            vec![AnnotationSegment {
+                node_id,
+                range: Range {
+                    start: 110,
+                    end: 120,
+                },
+                strand: Strand::Reverse,
+            }]
+        );
+    }
+}
